requirements.txt
smbus2==0.4.3
RPi.GPIO==0.7.1
python-socketio==5.11.3
eventlet==0.36.1
python-dotenv==1.0.1
celery==5.4.0
redis==5.0.7
gpsd-py3==0.3.0
skeleton-xml @ git+https://github.com/KostikShutov/skeleton-xml@main
